requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.5.0
async-timeout==4.0.3
attrs==24.2.0
blinker==1.8.2
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
cors==1.0.1
distro==1.9.0
exceptiongroup==1.2.2
filelock==3.16.1
flask==3.0.3
Flask-Cors==5.0.0
frozenlist==1.4.1
future==1.0.0
gevent==24.2.1
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
importlib-metadata==8.5.0
itsdangerous==2.2.0
jinja2==3.1.4
jiter==0.6.1
MarkupSafe==2.1.5
multidict==6.1.0
openai==1.51.2
propcache==0.2.0
pycparser==2.22
pydantic==2.9.2
pydantic-core==2.23.4
PySocks==1.7.1
python-dotenv==1.0.1
requests==2.32.3
requests-file==2.1.0
sniffio==1.3.1
tldextract==5.1.2
tqdm==4.66.5
typing-extensions==4.12.2
urllib3==2.2.3
werkzeug==3.0.4
yarl==1.15.0
zipp==3.20.2
zope.event==5.0
zope.interface==7.1.0